requirements.txt
streamlit
torch
transformers
numpy
Pillow 